renaming, restructure

+++ packages/npm/jns42-core/src/foreign/wrapper.ts
@@ -1,12 +1,12 @@
 import assert from "assert";
 import { NULL_POINTER, Pointer } from "../utils/index.js";
 
-export abstract class Wrapper {
+export abstract class ForeignObject {
   private disposed = false;
   private owned = true;
 
   constructor(public readonly pointer: Pointer) {
-    assert(pointer !== NULL_POINTER);
+    //
   }
 
   public letGo() {
@@ -18,7 +18,9 @@ export abstract class Wrapper {
   [Symbol.dispose]() {
     assert.equal(this.disposed, false);
     if (this.owned) {
-      this.drop();
+      if (this.pointer !== NULL_POINTER) {
+        this.drop();
+      }
     }
     this.disposed = true;
   }
